улучшена горячая отладка.

+++ mss_example/SoftConsole/mss_top_sb_MSS_CM3/mss_top_sb_MSS_CM3_app/Modules/micros/micros.c
@@ -22,9 +22,9 @@ void Timer1_IRQHandler()
 	t_h += 10;
 	MSS_TIM1_clear_irq();
 }
-int16_t my_atoi (uint8_t *buf, uint8_t l)
+int32_t my_atoi (uint8_t *buf, uint8_t l)
 {
-	int16_t tmp=0;
+	int32_t tmp=0;
 	uint8_t i =0;
 	for(i=0; i<l;i++)
 	{
